requirements.txt
alchemy==20.5
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
bcrypt==4.0.1
certifi==2023.7.22
cffi==2.0.0
chardet==3.0.4
charset-normalizer==3.4.4
click==8.1.8
contourpy==1.3.0
cryptography==46.0.3
cycler==0.10.0
defusedxml==0.7.1
dnspython==2.7.0
docutils==0.22.4
ecdsa==0.19.1
email-validator==2.3.0
exceptiongroup==1.3.1
fastapi==0.109.0
filelock==3.0.12
filetype==1.2.0
fonttools==4.60.2
h11==0.16.0
idna==2.8
importlib_resources==6.5.2
kiwisolver==1.4.7
lockfile==0.12.2
matplotlib==3.9.4
numpy==2.0.2
opencv-python==4.12.0.88
opencv-python-headless==4.8.0.74
packaging==25.0
passlib==1.7.4
pi_heif==1.1.1
pillow==11.3.0
pillow-avif-plugin==1.5.2
pillow_heif==1.1.1
pyasn1==0.6.1
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.10.1
pyparsing==2.4.7
python-daemon==2.2.4
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.5.0
python-magic==0.4.27
python-multipart==0.0.20
PyYAML==6.0.3
requests==2.22.0
requests-toolbelt==1.0.0
roboflow==1.1.9
rsa==4.9.1
scipy==1.13.1
six==1.17.0
SQLAlchemy==2.0.45
starlette==0.35.1
supervision==0.27.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==1.25.11
uvicorn==0.27.0
zipp==3.23.0
